+++ packages/manager/src/features/Dashboard/Dashboard.tsx
@@ -17,15 +17,19 @@ import {
 import Typography from 'src/components/core/Typography';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import Grid from 'src/components/Grid';
+import MaintenanceBanner from 'src/components/MaintenanceBanner';
+import TaxBanner from 'src/components/TaxBanner';
 import TagImportDrawer from 'src/features/TagImport';
 import useFlags from 'src/hooks/useFlags';
 import { handleOpen } from 'src/store/backupDrawer';
+import { addNotificationsToLinodes } from 'src/store/linodes/linodes.helpers';
 import getEntitiesWithGroupsToImport, {
   emptyGroupedEntities,
   GroupedEntitiesForImport
 } from 'src/store/selectors/getEntitiesWithGroupsToImport';
 import { openDrawer as openGroupDrawer } from 'src/store/tagImportDrawer';
 import { MapState } from 'src/store/types';
+import { formatNotifications } from 'src/utilities/formatNotifications';
 import shouldDisplayGroupImport from 'src/utilities/shouldDisplayGroupImportCTA';
 import { storage } from 'src/utilities/storage';
 import BackupsDashboardCard from './BackupsDashboardCard';
@@ -38,9 +42,6 @@ import NodeBalancersDashboardCard from './NodeBalancersDashboardCard';
 import PromotionsBanner from './PromotionsBanner';
 import TransferDashboardCard from './TransferDashboardCard';
 import VolumesDashboardCard from './VolumesDashboardCard';
-
-import MaintenanceBanner from 'src/components/MaintenanceBanner';
-import TaxBanner from 'src/components/TaxBanner';
 
 type ClassNames = 'root';
 
@@ -151,7 +152,13 @@ export const Dashboard: React.StatelessComponent<CombinedProps> = props => {
 };
 
 const mapStateToProps: MapState<StateProps, {}> = (state, ownProps) => {
-  const linodesData = state.__resources.linodes.entities;
+  const linodes = state.__resources.linodes.entities;
+  const notifications = state.__resources.notifications.data || [];
+
+  const linodesWithMaintenance = addNotificationsToLinodes(
+    formatNotifications(notifications),
+    linodes
+  );
 
   return {
     accountBackups: pathOr(
@@ -163,10 +170,12 @@ const mapStateToProps: MapState<StateProps, {}> = (state, ownProps) => {
     userTimezone: pathOr('', ['data', 'timezone'], state.__resources.profile),
     userTimezoneLoading: state.__resources.profile.loading,
     userTimezoneError: path(['read'], state.__resources.profile.error),
-    someLinodesHaveScheduledMaintenance: linodesData
-      ? linodesData.some(eachLinode => !!eachLinode.maintenance)
+    someLinodesHaveScheduledMaintenance: linodesWithMaintenance
+      ? linodesWithMaintenance.some(eachLinode => !!eachLinode.maintenance)
       : false,
-    linodesWithoutBackups: linodesData.filter(l => !l.backups.enabled),
+    linodesWithoutBackups: linodesWithMaintenance.filter(
+      l => !l.backups.enabled
+    ),
     managed: pathOr(
       false,
       ['__resources', 'accountSettings', 'data', 'managed'],
