+++ packages/manager/src/store/linodes/linodes.helpers.ts
@@ -40,20 +40,20 @@ export const addNotificationsToLinodes = (
 
     return foundNotification
       ? {
-        ...eachLinode,
-        maintenance: {
-          /**
-           * "when" and "until" are not guaranteed to exist
-           * if we have a maintenance notification
-           */
-          when: foundNotification.when,
-          until: foundNotification.until,
-          type: foundNotification.label as Type
+          ...eachLinode,
+          maintenance: {
+            /**
+             * "when" and "until" are not guaranteed to exist
+             * if we have a maintenance notification
+             */
+            when: foundNotification.when,
+            until: foundNotification.until,
+            type: foundNotification.label as Type
+          }
         }
-      }
       : {
-        ...eachLinode,
-        maintenance: null
-      };
+          ...eachLinode,
+          maintenance: null
+        };
   });
 };
