+++ packages/manager/src/store/linodes/linodes.reducer.ts
@@ -4,7 +4,6 @@ import updateById from 'src/utilities/updateById';
 import updateOrAdd from 'src/utilities/updateOrAdd';
 import { isType } from 'typescript-fsa';
 import {
-  addNotificationsToLinodes,
   createLinodeActions,
   deleteLinode,
   deleteLinodeActions,
@@ -15,7 +14,6 @@ import {
   upsertLinode
 } from './linodes.actions';
 
-import { addNotificationsToLinodes as _addNotificationsToLinodes } from './linodes.helpers';
 import { LinodeWithMaintenanceAndMostRecentBackup } from './types';
 
 const getId = <E extends HasNumericID>({ id }: E) => id;
@@ -151,18 +149,6 @@ const reducer: Reducer<State> = (state = defaultState, action) => {
       params: { linodeId }
     } = action.payload;
     const entities = state.entities.filter(({ id }) => id !== linodeId);
-
-    return {
-      ...state,
-      entities,
-      results: entities.map(getId)
-    };
-  }
-
-  if (isType(action, addNotificationsToLinodes)) {
-    const { payload: notifications } = action;
-
-    const entities = _addNotificationsToLinodes(notifications, state.entities);
 
     return {
       ...state,
